Implement Search feature
This enables the user to lookup a city, by entering the name.

Input is debounced (250ms), then the api is queried, if any matches
are received, they're displayed in a FlatList.

An ActivityIndicator provides user feedback while waiting for the api
to reply.

The api response codes are not documented in their site and hits are
limited in the free-tier, so further analysis should eventually be
performed.

Ref #2