Implement logic to add city based on user selection
Each result from the search screen is now touchable.

When the user taps on a city, the following takes place:

- The search TextInput is cleared
- The ADD_CITY action is dispatched
- The app navigates to the City screen

The model reducer adds the city to its sub state tree, taking into
account the 5 cities limit.

If this is the 6th city to be added, we drop the first from the list.

The env reducer uses the ADD_CITY action to reset the list of potential
cities (the list in the Search screen).

Currently, the navigation simply stacks City screen on top of Search,
but it could be possible to reset the navigation state, so that City
is stacked on top of Cities (the 'Home' screen).

Ref #3