Changes to better support testing
In connected components, export the base component (not connected), so
that it can be tested without re-testing connect.

This is possible, because in mapStateToProps, state is accessed via
selectors, which can be tested separately.

Also, in mapDispatchToProps, actions are defined by action creators
that can also be tested separately.

Some discussions on this matter:
https://github.com/reactjs/react-redux/issues/325#issuecomment-199449298
https://github.com/reactjs/redux/issues/1534
https://jsramblings.com/2018/01/15/3-ways-to-test-mapStateToProps-and-mapDispatchToProps.html

Ref #14