Convert city id from number to string
Having the id as number means multiple toString conversion, since
libraries (specially react-native-swipe-list), want string keys.

A Map wouldn't really solve the issue either, since we'd still need to
convert.

So, as soon as we receive a reply from the api endpoint, we convert the
id to string, so our store and components only ever see a string id.

Ref #15